+++ scripts/generate-tmlanguage.mts
@@ -14,6 +14,7 @@ import * as fs from "fs/promises";
 import * as tokens from "../packages/language/src/parser/tokens.js";
 import * as sqlTokens from "../packages/language/src/parser/tokens/sql-tokens.js";
 import * as cicsTokens from "../packages/language/src/parser/tokens/cics-tokens.generated.js";
+import { toPattern } from "./pattern-trie.js";
 
 const manual = JSON.parse(
   await fs.readFile(
@@ -38,99 +39,6 @@ for (const [text, type] of mergedMap.entries()) {
   } else {
     storageKeywords.push(lowerText);
   }
-}
-
-// Unnecessary (?) optimization to avoid generating a huge regex with each keyword as an alternative
-// Instead, we serialize a trie structure into the regex, which should be more efficient to execute
-// Note: The regex generated by this step is completely unreadable.
-function longestCommonPrefix(words: string[]): string {
-  if (words.length === 0) return "";
-  let prefix = words[0];
-  for (let i = 1; i < words.length; i++) {
-    let j = 0;
-    while (
-      j < prefix.length &&
-      j < words[i].length &&
-      prefix[j] === words[i][j]
-    ) {
-      j++;
-    }
-    prefix = prefix.slice(0, j);
-    if (!prefix) {
-      break;
-    }
-  }
-  return prefix;
-}
-
-function isSingleAtom(pattern: string): boolean {
-  if (pattern.length === 1) {
-    return true;
-  } else if (pattern.length === 2 && pattern[0] === "\\") {
-    return true;
-  } else if (pattern.startsWith("(?:") && pattern.endsWith(")")) {
-    return true;
-  }
-  return false;
-}
-
-function buildAlternation(words: string[]): string {
-  if (words.length === 1) {
-    return words[0]
-      .split("")
-      .map((c) => tokens.escapeRegExp(c))
-      .join("");
-  }
-
-  // Group by first character
-  const groups = new Map<string, string[]>();
-  for (const word of words) {
-    const first = word[0];
-    if (!groups.has(first)) groups.set(first, []);
-    groups.get(first)!.push(word);
-  }
-
-  const parts: string[] = [];
-  for (const group of groups.values()) {
-    const prefix = longestCommonPrefix(group);
-    const escapedPrefix = prefix
-      .split("")
-      .map((c) => tokens.escapeRegExp(c))
-      .join("");
-    const suffixes = group.map((w) => w.slice(prefix.length));
-    const hasEmpty = suffixes.some((s) => s === "");
-    const nonEmpty = suffixes.filter((s) => s !== "");
-
-    if (nonEmpty.length === 0) {
-      parts.push(escapedPrefix);
-    } else {
-      const inner = buildAlternation(nonEmpty);
-      const suffix = hasEmpty
-        ? isSingleAtom(inner)
-          ? `${inner}?`
-          : `(?:${inner})?`
-        : inner;
-      parts.push(escapedPrefix + suffix);
-    }
-  }
-
-  if (parts.length === 1) return parts[0];
-  return `(?:${parts.join("|")})`;
-}
-
-function toPattern(keywords: string[]) {
-  const sorted = [...new Set(keywords)].sort();
-
-  // Group by first character so the top-level join doesn't need an outer (?:...)
-  const groups = new Map<string, string[]>();
-  for (const kw of sorted) {
-    if (!kw) continue;
-    if (!groups.has(kw[0])) groups.set(kw[0], []);
-    groups.get(kw[0])!.push(kw);
-  }
-
-  const parts = [...groups.values()].map(buildAlternation);
-  return `(?i)\\b(?:${parts.join("|")})\\b`;
 }
 
 const controlPattern = toPattern(controlKeywords);
